Settings to vx (#78)
* UWP converted create account

* Droid VxCreateAccount

* iOS create account vx

* Remove the ConfirmPassword

* Started adding Login

* iOS improve buttons

* iOS replace reset password

* Android removed login views

* UWP remove login views

* Add hot reload for UWP

* Replaced ConnectAccount and ExistingUser

* Rewrite MyAccount and DeleteAccount

* Add sound effect settings

* Fix Droid compiling

* Rewrote tons of settings views

* Auto-focus UWP password box

* Replace change username and password

* Convert ChangeEmail

* Add Get/SetState and propvx snippet

* Switch to VxValue

* Fix up Droid and iOS

* Rewrite even more settings views

* Fix Droid compiling

* iOS clean up namespaces

+++ Vx/Views/View.cs
@@ -40,8 +40,8 @@ namespace Vx.Views
         /// </summary>
         public float Opacity { get; set; } = 1;
 
-        public double Width { get; set; } = double.NaN;
-        public double Height { get; set; } = double.NaN;
+        public float Width { get; set; } = float.NaN;
+        public float Height { get; set; } = float.NaN;
 
         public Action<View> ViewRef { get; set; }
